[hooks] decouple hook chain from clsx/tailwind
The hook handler (`src/hooks/handler.ts`) was transitively pulling in
`lib/utils.ts` via:

  handler.ts
    -> resolve-permission-mode.ts
      -> lib/codex-sessions.ts
        -> lib/log-entries.ts
          -> lib/utils.ts (imports clsx + tailwind-merge)

`lib/utils.ts` is the dashboard's shadcn `cn()` helper — it has no business
on the runtime path of a CLI hook. The only thing the hook chain actually
needed from it was `formatDate`, an `Intl.DateTimeFormat` wrapper with no
third-party deps.

When a Claude Code session opened a clone of this repo whose
`$CLAUDE_PROJECT_DIR` resolved into an ancestor that shadowed bun's global
install cache (e.g. `/home/nivedit/prs/failproofai-231/`, with `node_modules`
missing locally but `/home/nivedit/node_modules` present and clsx-less),
every hook event crashed with:

  ResolveMessage: Cannot find package 'clsx' from .../lib/utils.ts

Move `formatDate` into a dedicated `lib/format-date.ts` and update the four
server-side and three dashboard callers. `lib/utils.ts` now only exports
`cn` so the hook handler no longer transitively requires `clsx`.
