+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 ### Fixes
+- Decouple the hook handler from `lib/utils.ts` so `--hook <Event>` no longer crashes with `Cannot find package 'clsx'` when launched from a clone that hasn't run `bun install`. The handler's static graph (`handler.ts` → `resolve-permission-mode.ts` → `lib/codex-sessions.ts` → `lib/log-entries.ts`) was reaching `lib/utils.ts` just to use `formatDate`, which dragged in the dashboard's `clsx` + `tailwind-merge` deps. Move `formatDate` into its own `lib/format-date.ts` (no third-party deps) and update server-side callers (`lib/log-entries.ts`, `lib/projects.ts`) plus dashboard callers (`app/components/{project-list,sessions-list}.tsx`, `app/project/[name]/page.tsx`) and the `__tests__/lib/utils.test.ts` import. `lib/utils.ts` now only exports `cn`.
 - Fix `mintlify validate` failing on `docs/de/dashboard.mdx` after #228 added German „..." quotation marks inside two `<Tab title="…">` attributes — the inner straight `"` ended the JSX attribute value, then the leftover `"` produced an `Unexpected character "` parse error. Drop the inner quotes (matching how every other locale renders the tab labels) so the docs CI job passes again.
 - Fix `block-read-outside-cwd` falsely denying Bash commands that contain unquoted glob patterns or compound argv tokens. The unquoted-path regex in `extractAbsolutePaths` (`src/hooks/builtin-policies.ts`) only excluded `[a-zA-Z0-9_.\-~\\]` from its negative lookbehind, so a `/`-prefixed glob suffix or volume-mount target was misread as a standalone absolute path: `grep ... docs/*/dashboard.mdx | head` extracted `/dashboard.mdx`, and `docker run -v /host:/docs ...` extracted `/docs`. Both paths then resolved outside cwd and the read-like-command branch denied the call. Add `*?:=` to the lookbehind exclusion class so a `/` immediately following a glob meta or a separator no longer starts a match. Existing whitelisted/quoted/space-separated path detection is unchanged.
